service_a_auth/requirements.txt
flask
flask-cors
flask-mail
bcrypt
python-dotenv
gunicorn